+++ pkg/R/graphs.R
@@ -14,14 +14,9 @@
 
 ### Determine the matrix of neighbourhood orders
 ### given the binary matrix of first-order neighbours.
-### Working horse: spdep::nblag()
 
 nbOrder <- function (neighbourhood, maxlag = 1)
 {
-    if (!requireNamespace("spdep"))
-        stop("package ", sQuote("spdep"),
-             " is required to determine neighbourhood orders")
-
     stopifnot(isScalar(maxlag), maxlag > 0)
     checkNeighbourhood(neighbourhood)
     neighbourhood <- neighbourhood == 1           # convert to binary matrix
@@ -33,35 +28,33 @@ nbOrder <- function (neighbourhood, maxlag = 1)
         return(neighbourhood)
     }
 
-    ## convert binary adjacency matrix to spdep's "nb" class, "manually",
-    ## mimicing spdep::mat2listw(neighbourhood)$neighbours; there is no mat2nb()
-    nb <- structure(
-        apply(neighbourhood, 1L, function(x)
-            if (length(nbs <- which(x))) nbs else 0L,
-            simplify = FALSE),
-        class = "nb")
-
-    ## compute higher order neighbours using spdep::nblag()
-    nb.lags <- spdep::nblag(nb, maxlag=maxlag)
+    ## list of indexes of first-order neighbours by region
+    first <- apply(neighbourhood, 1L, which, simplify = FALSE)
 
     ## Side note: fast method to determine neighbours _up to_ specific order:
     ## crossprod(neighbourhood) > 0  # up to second order neighbours (+set diag to 0)
     ## (neighbourhood %*% neighbourhood %*% neighbourhood) > 0  # up to order 3
     ## and so on...
 
-    ## convert to a single matrix
-    nbmat <- neighbourhood   # logical first-order matrix
-    storage.mode(nbmat) <- "numeric"
-    for (lag in 2:maxlag) {
-        if (any(spdep::card(nb.lags[[lag]]) > 0L)) { # any neighbours of this order
-            nbmat.lag <- spdep::nb2mat(nb.lags[[lag]], style="B",
-                                       zero.policy=TRUE)
-            nbmat <- nbmat + lag * nbmat.lag
+    ## now find recursive neighbours for each region
+    nbmat <- `diag<-`(neighbourhood, NA)  # skip self in which() below
+    for (i in seq_len(nregions)) {
+        nblags <- as.integer(nbmat[i,])
+        lag <- 1L
+        while (lag < maxlag) {
+            nbs <- which(nblags == lag)
+            nbs2 <- unlist(first[nbs])
+            new <- intersect(nbs2, which(nblags == 0))
+            if (length(new)) {
+                lag <- lag + 1L
+                nblags[new] <- lag
+            } else break
         }
+        nbmat[i,] <- nblags
     }
-    attr(nbmat, "call") <- NULL
-    storage.mode(nbmat) <- "integer"
+    diag(nbmat) <- 0L
 
+    ## FIXME: drop this annoying message
     ## message about maximum neighbour order by region
     maxlagbyrow <- apply(nbmat, 1, max)
     message("Note: range of maximum neighbour order by region is ",